+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## 0.1.0
 
+- HTTP to HTTPS redirect by design
 - Added HTTPS part and threading
 - Headers dump
 - Small fixes to proper static content handling